backend/requirements.txt
fastapi==0.124.4
uvicorn[standard]==0.38.0
pre-commit==4.3.0
httpx==0.28.1
asyncpg==0.31.0
python-dotenv==1.2.1
python-jose==3.5.0
passlib[bcrypt]==1.7.4
email-validator==2.3.0
